Rename Enrollment CheckIn components and paths
