+++ src/views/EnrollmentCheckIn/components/CompletedCheckIn/index.js
@@ -2,7 +2,7 @@ import { Button, Grid, Typography } from '@material-ui/core';
 import { Link } from 'react-router-dom';
 import { gordonColors } from 'theme';
 // @TODO CSSMODULES - outside directory
-import styles from '../../AcademicCheckIn.module.css';
+import styles from '../../EnrollmentCheckIn.module.css';
 
 const CompletedCheckIn = ({ basicInfo }) => {
   const cyan = gordonColors.primary.cyan;
