+++ src/views/EnrollmentCheckIn/components/EnrollmentCheckInWelcome/index.js
@@ -1,7 +1,7 @@
 import { Grid, Typography } from '@material-ui/core';
 import { gordonColors } from 'theme';
 // @TODO CSSMODULES - outside directory
-import styles from '../../AcademicCheckIn.module.css';
+import styles from '../../EnrollmentCheckIn.module.css';
 
 const EnrollmentCheckInWelcome = ({ basicInfo, hasMajorHold, holds }) => {
   const blue = gordonColors.primary.blue;
